drivers/media/video build fix for modular builds
testing of the -tip tree found the following missing symbols build bug:

    LD      init/built-in.o
    LD      .tmp_vmlinux1
    drivers/built-in.o: In function `vidioc_g_ctrl':
    radio-maxiradio.c:(.text+0xd3337): undefined reference to `video_devdata'
    drivers/built-in.o: In function `vidioc_g_frequency':
    radio-maxiradio.c:(.text+0xd3368): undefined reference to `video_devdata'
    drivers/built-in.o: In function `vidioc_s_ctrl':
    [...]

which happens if certain media drivers are modular:

    http://redhat.com/~mingo/misc/config-Thu_May_15_18_10_27_CEST_2008.bad

the problem is:

    CONFIG_VIDEO_MEDIA=m
    CONFIG_VIDEO_DEV=y
    CONFIG_RADIO_MAXIRADIO=y
    CONFIG_VIDEO_V4L2=y

VIDEO_MEDIA went modular and took a bunch of symbols with it.

I went for the more robust fix in the kbuild space instead of further
uglifying the Kconfig space.  Adding a "depends on VIDEO_DEV" to
VIDEO_MEDIA does not solve the problem.

This problem too is a side-effect of all the uglifying effects of the
well-known kconfig tool bug: it does not properly map out dependencies
between components and thus allows nonsensical results that introduces
obscure build bugs with certain build combinations.

Signed-off-by: Ingo Molnar <mingo@elte.hu>