kmemcheck: fix 64-bit address lookup
There was actually a bug where the address passed into address_get_pte()
was truncated to 32-bit on 64-bit causing 64-bit to hang as soon as the
first tracked memory was being accessed, since the page fault was
apparently handled, but not really.

With this patch in place, kmemcheck on 64-bit "works" in that the #PF/#DB
sequence works smoothly; there are no hangs or unhandled page faults,
though the opcode decoding is wrong and will lead to many false positives.

Signed-off-by: Vegard Nossum <vegard.nossum@gmail.com>
Signed-off-by: Ingo Molnar <mingo@elte.hu>