kprobes: fix ARM compilation break due to kprobe_mutex
Compilation of kprobes on arm (as of 2.6.26-rc2) breaks due to:
http://lkml.org/lkml/2008/4/9/159

This fix is a logical extension of the said commit to remove
the use of kprobe_mutex from arch/arm/kernel/kprobes.c.

Mathieu muttered:
Yes, it looks like when I created this patch, arm did not have kprobes
support yet. This fix is therefore correct.

Signed-off-by: Abhishek Sagar <sagar.abhishek@gmail.com>
Acked-by: Mathieu Desnoyers <mathieu.desnoyers@polymtl.ca>
Signed-off-by: Thomas Gleixner <tglx@linutronix.de>