x86, I/O APIC: timer through 8259A revamp
There is some seemingly unnecessary cruft accumulated in the 8254 timer
setup path for the I/O APIC.  Some attempts have been made to cater for
broken BIOSes reporting the ExtINTA I/O APIC pin cascaded from the
master 8259A as the native timer interrupt input, which interfere with
the setup sequence in a non-obvious way and require a pair of additional
kernel command-line parameters.

I think this is unnecessary.  All the bits are already in place.  With
the removal of AEOI bits for non-i82489DX systems the whole setup can be
simplified.  Here is a set of changes that reuse the existing code for
potentially broken systems.  The observation is broken systems do not
have a notion of ExtINTA pins, so only one pin/apic pair will be
determined as useful for the timer interrupt.  Therefore it should be
safe to select this single pin/apic to test both directly and through
the 8259A.

This change implements this approach by copying timer pin information
found to the other pin if one has been found only.  This way both tests
are always performed.  The timer input of the 8259A is carefully
unmasked if needed only and masked again if found unneeded.  This is
because some systems seem sensitive for the master 8259A interrupt being
active even if all the APIC inputs this line is wired to are masked.

The "disable_8254_timer" and "enable_8254_timer" kernel parameters are
removed as no longer needed.

Signed-off-by: Maciej W. Rozycki <macro@linux-mips.org>
Signed-off-by: Ingo Molnar <mingo@elte.hu>