futex, 64-bit support: generic code
This patch adds the architecture independent changes.  For this it has to
adjust the futex_atomic_op_inuser function in arch-specific code but this
in no real code change.  That function is inlined for most architectures.

The real changes depend on the definition of a single type: futex_val_t.
If the arcitecture defines this type to something at least as large as u64
then all the 64-bit futex code is automagically enabled.  If the type is
not defined sufficiently using the FUTEX_64_FLAG flag triggers an ENOSYS
error.

We already have a flag supported for futexes, the private flag.  When it
was introduced a pointer to a semaphore for the memory handling was passed
around.  I changed that.  Now the flags are passed in a word and the
semaphore is located in every function.  This might have added a minute
fraction of code but

- there is no runtime difference for this since the semaphore address
  is still only computed once

- the compiler now has more insight into when no semaphore is used and
  it can perform more optimization (see the futex_lock_mm definition).

- the restart handling is cleaned up a bit as well, as a consequence

The overall code size increase is minute, as can be guessed from the diffstat
numbers below.  Must of the changes are changes to existing code and not
new code.

Signed-off-by: Ulrich Drepper <drepper@redhat.com>
Signed-off-by: Ingo Molnar <mingo@elte.hu>