futex, 64-bit support: x86 architecture support, 64-bit
This patch adds 64-bit futex support for x86-64.  As can be seen, it consists
almost exclusively of the enablement for the FUTEX_WAKE_OP operation in the
form of new inline assembly code.  The magic to enable the 64-bit support is
the one little type definition in the types.h file.

Signed-off-by: Ulrich Drepper <drepper@redhat.com>
Signed-off-by: Ingo Molnar <mingo@elte.hu>