ACPI, x86: disable the C2C3_FFH access mode if MWAIT is unsupported
After the following commit is merged, C2C3_FFH mode will be enabled and
passed into the _PDC object. But because MWAIT is unsupported on some laptops,
the different power states will be returned by the _CST object.
Maybe it will be better to disable C2C3_FFH access mode if MWAIT is unsupported.

commit 991528d7348667924176f3e29addea0675298944
Author: Venkatesh Pallipadi <venkatesh.pallipadi@intel.com>
Date:   Mon Sep 25 16:28:13 2006 -0700

    ACPI: Processor native C-states using MWAIT

http://bugzilla.kernel.org/show_bug.cgi?id=10807

Signed-off-by: Zhao Yakui <yakui.zhao@intel.com>
Signed-off-by: Li Shaohua <shaohua.li@intel.com>
Signed-off-by: Ingo Molnar <mingo@elte.hu>