x86, 64-bit: standardize printk_address()
Changes the symbol-in-module format from

	:ext3:add_dirent_to_buf+0x6c/0x26c

to

	add_dirent_to_buf+0x6c/0x26c [ext3]

the latter was the standard format anyway (it's what "sprint_symbol()"
gives you), and traps_64.c was the odd man out.

In fact, traps_32.c already uses the standard print_symbol() format, so
it really was an issue of the 64-bit code being odd (and I assume that
this also means that it cannot break the oops tracker, since it already
had to be able to handle both formats).

I also removed the KALLSYMS dependency, so if KALLSYMS isn't enabled it
will now give the same hex format twice, but I doubt we really care
(such stack traces are unreadable whether it shows up once or twice, and
the simplicity is worth it).

If people do just a few more conversions like this, then the 52 added
lines in lib/vsnprintf.c are more than made up for by removed lines
elsewhere (and more readable source code).

Signed-off-by: Ingo Molnar <mingo@elte.hu>