printk: add support for '%pS'
I've cleaned up the lib/vsnprintf.c patch to fix up some minor
details. For example, it should not default to a field width of
"2*sizeof(unsigned long)" - it should do that only if it ends up printing
the pointer as a hex number.

(the "precision" thing was moved into the 'pointer()' function, and
 to after the special case handling).

Signed-off-by: Linus Torvalds <torvalds@linux-foundation.org>
Signed-off-by: Ingo Molnar <mingo@elte.hu>