kmemcheck: mark SMP support BROKEN
It was recently discovered that our SMP support is in fact broken; if
the NMI watchdog is enabled, the machine will hang at boot, with the
message "Testing NMI watchdog".

Most likely we're getting an extra NMI (or missing one) and as a
result we might be waiting for a "phantom" CPU, i.e. one that doesn't
exist in the first place, because we bumped the counter one time too
many or one time too few.

Signed-off-by: Vegard Nossum <vegard.nossum@gmail.com>
Signed-off-by: Ingo Molnar <mingo@elte.hu>