fastboot: create a "asynchronous" initlevel
This patch creates an asynchronous initlevel (6a) which is at the same
level as the normal device initcalls, but with the difference that they
are run asynchronous from all the other initcalls. The purpose of this
*selective* level is that we can move long waiting inits that are not
boot-critical to this level one at a time.

To keep things not totally insane, the asynchronous initcalls are async
to the other initcalls, but are still ordered to themselves; think of it
as "bottom-half-not-softirq". This has the benefit that async drivers
still have stable device ordering between them.

Signed-off-by: Arjan van de Ven <arjan@linux.intel.com>
Signed-off-by: Ingo Molnar <mingo@elte.hu>