iwlwifi: fix build bug in "iwlwifi: fix LED stall"
-tip testing found the following build failure:

  drivers/net/wireless/iwlwifi/iwl-led.c: In function ‘iwl_led_brightness_set’:
  drivers/net/wireless/iwlwifi/iwl-led.c:198: error: ‘led_type_str’ undeclared (first use in this function)
  drivers/net/wireless/iwlwifi/iwl-led.c:198: error: (Each undeclared identifier is reported only once
  drivers/net/wireless/iwlwifi/iwl-led.c:198: error: for each function it appears in.)

Triggered if this driver is built with !CONFIG_IWLWIFI_DEBUG. Introduced
by commit 0eee61273.

The best fix is to make led_type_str available as a zero-size symbol and to
only add members to the array if CONFIG_IWLWIFI_DEBUG is set. This way
there's no overhead in the debugging case and we have type checking in the
IWL_DEBUG_LED() macro as well.

Signed-off-by: Ingo Molnar <mingo@elte.hu>