ftrace: do not trace NMI contexts
Dynamic ftrace can not trace the nmi code. The reason is
that it is not safe to modify code text on an SMP system when
the other CPUS are exectuting code.

Unfortunately, there's no uniform API to disable NMIs on other
CPUS - nor does one look easy (or even desirable).

So warn about NMI entries. This warning can be used to mark
functions notrace that can be called in an NMI context.

Signed-off-by: Steven Rostedt <srostedt@redhat.com>
Signed-off-by: Ingo Molnar <mingo@elte.hu>