kmemcheck: implement REP MOVS/STOS emulation
It turns out that certains CPUs (so far we've seen the pattern only
in P4 machines) will single-step a REP MOVS/STOS instruction only
after RCX/ECX/CX has reached zero. This means that prior to this
patch, kmemcheck would only detect the very first such instruction
on those machines.

Note that only REP MOVS/STOS instructions seem to be affected by this;
REP CMPS, REP LODS, etc. are fine.

It's NOT possible to do a "real" emulation of the instruction. We
might access a kernel address in %esi and a userspace address in %edi.
And if we try to access userspace from kmemcheck (which is called from

My final hack turned out like this (I'm quite proud of it): Instead of
emulating the _whole_ REP MOVS/STOS, we only emulate the REP part.
That is, on #PF, we increment %eip by one, which means that when the
give is the #DB straight afterwards). Now, in the #DB, we check the
flag that says "was this really a REP instruction?" and if it was, we
start counting down %ecx and rewinding %eip each time until %ecx is 0.
Each time we return to the original instruction and let the CPU
execute it natively. When %ecx is 0, we turn off single-stepping and
hide the pages again.

What's also neat about this, although it sounds slow, is that we
actually only modify the page tables at the beginning and the end of a
whole REP MOVS/STOS sequence, so this should in theory be faster on
non-P4s as well (we skip TWO TLB flushes and don't need the #PF
in-between each iteration of the loop either).

Signed-off-by: Vegard Nossum <vegardno@ifi.uio.no>