kmemcheck: hide/show pages in each iteration of a REP instruction
In certain cases when the REP MOVS/STOS instruction crossed a page
boundary, the previous page would not be hidden. Thus, subsequent
accesses to that page would not be detected, and we would fail to
update the shadow state. However, if another multi-operand instruction
happened to get a page fault (with this page in its other operand),
kmemcheck would notice that this page is ours and hide it again
afterwards. This would lead to correct execution, except for the
possibly huge number of false-positive warnings that were reported
because we failed to update the shadow state.

So for now, we take the slow, but safe route of hiding/showing the
pages in question on each iteration of the REP instruction.

Signed-off-by: Vegard Nossum <vegard.nossum@gmail.com>