kmemcheck: rip out the optimized memset()
Although a good idea in principle, the task is really too complex to
be implemented in an elegant manner. It's too much code to keep around,
so let's just drop it to keep the code simple.

(Correctness is always more important and this was a premature
optimization to begin with.)

Signed-off-by: Vegard Nossum <vegard.nossum@gmail.com>