kmemcheck: hide/show pages in each iteration of a REP instruction #2
I made a mistake when fixing the REP handling the first time around;
after the previous patch, we would rewind the instruction pointer
too far (back to the REP prefix), and we would get another page
fault.

This patch is admittedly not the most elegant I've seen, but it will
do the trick for now. Now we rewind only back to the instruction
proper, and single-step that over and over again, as was originally
intended.

Signed-off-by: Vegard Nossum <vegard.nossum@gmail.com>