Revert "kmemcheck: use set_memory_4k() instead of disabling PSE"
This reverts commit 10142e6c6778e201216f68bb378e139de7ee2551.

Unfortunately, set_memory_4k() does not work in atomic context (well,
when irqs are disabled), because it needs to flush TLBs on all CPUs.
While this shouldn't be a problem on uniprocessor, the pagetable API
does not distinguish between SMP and non-SMP and never allows
changing the page tables when irqs are disabled.

Conflicts:

	arch/x86/kernel/cpu/common.c
	arch/x86/kernel/kmemcheck.c
	arch/x86/mm/fault.c
	arch/x86/mm/init_32.c
	include/asm-x86/kmemcheck.h

Signed-off-by: Vegard Nossum <vegard.nossum@gmail.com>