include/asm-generic/bug.h - fix warning in mm/bootmem.c
fix this warning:

  mm/bootmem.c: In function ‘mark_bootmem’:
  mm/bootmem.c:321: warning: control reaches end of non-void function

This turns out to be a include/asm-generic/bug.h problem in the
!CONFIG_BUG case.

In that case we completely ignore BUG() sequences.

My first instinct was to use an infinite loop instead of a NOP, but that
creates a lot of extra code:

   text    data     bss     dec     hex filename
   8807935 2012144 5619712 16439791         fad9ef vmlinux.before
   8829752 2012144 5619712 16461608         fb2f28 vmlinux.after

about +0.25% in size.

So ... add a return value instead. We should not hit it in practice.

Signed-off-by: Ingo Molnar <mingo@elte.hu>