#ifdef complications in drivers/acpi/sleep/main.c
this warning:

  drivers/acpi/sleep/main.c:67: warning: ‘acpi_pm_disable_gpes’ defined but not used
  drivers/acpi/sleep/main.c:92: warning: ‘acpi_pm_prepare’ defined but not used
  drivers/acpi/sleep/main.c:107: warning: ‘acpi_pm_finish’ defined but not used
  drivers/acpi/sleep/main.c:128: warning: ‘acpi_pm_end’ defined but not used

Shows that this code has an identity crisis due to a maze of #ifdefs, in
the PM_SLEEP && !SUSPEND && !HIBERNATION case for example.

Instead of complicating the code with even more #ifdefs, one option
would be to convert these rather trivial wrappers to inline functions
(implemented in this commit). Maybe there's a better solution as well -
such as to reduce the many config options we have in this area?

No size difference with SUSPEND && HIBERNATION turned back on:

drivers/acpi/sleep/main.o:

   text	   data	    bss	    dec	    hex	filename
   2717	    976	     33	   3726	    e8e	main.o.before
   2717	    976	     33	   3726	    e8e	main.o.after

md5:
   44220906eff0ea6e1a3266b35dd82ac2  main.o.before.asm
   44220906eff0ea6e1a3266b35dd82ac2  main.o.after.asm