include/linux/init.h: fix warning in drivers/crypto/hifn_795x.c
fix this warning:

  drivers/crypto/hifn_795x.c:2729: warning: ‘hifn_remove’ defined but not used

this is caused by a weakness of the __devexit_p() macro in the
!CONFIG_HOTLPUG case - it returns NULL but keeps the function parameter
unused - triggering this warning.

So improve the __devexit_p() macro. Unfortunately, it looks a bit ugly,
because it's hard to create a structure initializer that 'uses' a parameter
while producing a NULL. Braced groups ({ }) cannot be used here.

[ Also clean up its surroundings while at it. ]

Signed-off-by: Ingo Molnar <mingo@elte.hu>