fix warning in net/netfilter/ipvs/ip_vs_ctl.c
Impact: fix potential kernel stack information exposure to user-space.

fix this warning:

  net/netfilter/ipvs/ip_vs_ctl.c: In function ‘ip_vs_genl_set_cmd’:
  net/netfilter/ipvs/ip_vs_ctl.c:3034: warning: ‘t.udp_timeout’ may be used uninitialized in this function
  net/netfilter/ipvs/ip_vs_ctl.c:3034: note: ‘t.udp_timeout’ was declared here

GCC seems to be correct here: in the !CONFIG_IP_VS_PROTO_UDP case
__ip_vs_get_timeouts keeps the timeout structure uninitialized. But
in a couple of places we use and copy this structure to user-space,
potentially exposing kernel-stack data to user-space.

While it's a minor issue that only triggers with certain config variations,
it's best to always initialize this structure in __ip_vs_get_timeouts().
(GCC will eliminate this memset)

Signed-off-by: Ingo Molnar <mingo@elte.hu>