fix warning in net/netfilter/nf_conntrack_proto_sctp.c
fix this warning:

  net/netfilter/nf_conntrack_proto_sctp.c: In function ‘sctp_packet’:
  net/netfilter/nf_conntrack_proto_sctp.c:376: warning: array subscript is above array bounds

GCC noticed that new_state could in theory be SCTP_CONNTRACK_MAX,
which is beyond the limit of the sctp_timeouts[] array.

This situation could happen if for_each_sctp_chunk() exits immediately.
That should not normally happen - but it's not obvious to me and it's
better to insert an extra check to the end of the loop for this case.

This gets rid of warning as well.

Signed-off-by: Ingo Molnar <mingo@elte.hu>