improve the structure of pnpbios_thread_init()
Impact: cleanup, size reduction. Possible (but unlikely) effect on PPC.

this warning:

  drivers/pnp/pnpbios/core.c: In function ‘pnpbios_thread_init’:
  drivers/pnp/pnpbios/core.c:578: warning: unused variable ‘task’

is caused because the 'task' variable is unused in the !CONFIG_HOTPLUG
case.

Instead of adding an #ifdef around the variable, simplify the code
instead. This is done by observing that the whole pnpbios_thread_init()
function only does something useful if CONFIG_HOTPLUG is enabled.

So make this whole kpnpbiosd sequence conditional cleanly: remove
the #ifdef from within the function and thus make it easier to read.
This also makes the kernel a bit smaller on !CONFIG_HOTPLUG, and
avoids the warning as well.

Caveat emptor: in theory check_legacy_ioport() call on PPC could
have side-effects - and we wont be calling it now. But it does not
seem to be a function that has side-effects - it just parses machine
descriptor nodes, so we should be fine.

Signed-off-by: Ingo Molnar <mingo@elte.hu>