kmemcheck (bitfields): don't actually initialize the bitfield
If we do a memset() on the bitfield, it will be cleared. What we
want is to mark it initialized, so do (only) that. This way, we
also don't change the behaviour of the code itself -- for example,
a cache constructor could decide to preserve a bitfield across
alloc/free, which would be bad for us, since we reset the actual
contents of the bitfield too. So fix that.

Signed-off-by: Vegard Nossum <vegard.nossum@gmail.com>