fix build bug in 2.6.30-rc4: xt_socket.c:(.text+0x57780): undefined reference to `nf_conntrack_untracked'
Not sure whether this has been reported already, but latest -git
triggers this build failure in -tip testing:

 net/built-in.o: In function `socket_mt':
 xt_socket.c:(.text+0x57780): undefined reference to `nf_conntrack_untracked'

i think it came in via the 3dacbdad2401c06b97d8d754974233a70c165536
networking tree merge.

The build failure is not particularly frequent (has a less than 1:50
chance so far and does not seem to affect any usual
allyes/allno/defconfigs), so i'd suggest to look for some
unanticipated .config combinations in the randconfig-generated
config below.

ah, i think it's due to:

 CONFIG_NF_CONNTRACK=m
 CONFIG_NETFILTER_XT_MATCH_SOCKET=y

conntrack a module, but xt-socket built-in.

my bet is on:

| commit 4b0706624930dc75c3b0d0df463d89759ef7de29
| Author: Laszlo Attila Toth <panther@balabit.hu>
| Date:   Fri Apr 24 16:55:25 2009 +0200
|
|     netfilter: Kconfig: TProxy doesn't depend on NF_CONNTRACK
|
|     Signed-off-by: Laszlo Attila Toth <panther@balabit.hu>
|     Signed-off-by: Patrick McHardy <kaber@trash.net>
causing this.

And indeed - reverting that commi solves the problem.

	Ingo

commit 310ef338a01ec7684bae8ba69ceb8cf64b5db65d
Author: Ingo Molnar <mingo@elte.hu>
Date:   Thu Apr 30 09:53:56 2009 +0200

    Revert "netfilter: Kconfig: TProxy doesn't depend on NF_CONNTRACK"

    This reverts commit 4b0706624930dc75c3b0d0df463d89759ef7de29.